MAINT: Renamed derived_operators -> operator_states
